virtio_net: remove hdr size from buf calculation
The number of virtio ring descriptors required for enqueueing is
determined by the packet data size plus the virtio header size.
The resulting descriptor count includes the virtio header size to be copied
to the ring.

If the difference between the source length (slen) and destination length
(dlen) is less than the size of the virtio header, we will end up skipping
the same amount of packet data during the DMA copy because of the enqueue
count calculation. Therefore, removed virtio header size subtraction
from the loop

Signed-off-by: Kommula Shiva Shankar <kshankar@marvell.com>
Change-Id: I7bcfa7323d5d209234704388efff2c1be33d155b
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/133739
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Nithin Kumar Dabilpuram <ndabilpuram@marvell.com>