lib/liquid_crypto: fix packet length in RSA API
Crypto operation such as RSA PKCS encrypt can be performed,
even for mod of 17 bytes, exponent of 1 byte and plain text
of 1 byte in length, in which case buf_len is 39 bytes,
including transport header.

Hence, allow smaller this data payload, but ensure pkt_len
in mbuf to be minimum packet size supported in transport.

Signed-off-by: Gowrishankar Muthukrishnan <gmuthukrishn@marvell.com>
Change-Id: Idd18eaa2daeb87f4138ae7a28a492ae34090562f
Reviewed-by: Anoob Joseph <anoobj@marvell.com>