app/card-mgr: application to manage the dao card
Application to manage dao card.
Based on arguments provided same application can run as client as well
as server.

In server mode, it receives the commands from client and process the
commands using gRPC APIs. Sends back the information to client.

In client mode, it receives the commands from user via cli and passes it
to server.

Signed-off-by: Vidya Sagar Velumuri <vvelumuri@marvell.com>
Change-Id: I837930e3e5fcb0a4c8a5491700462367e192f796
Reviewed-by: Anoob Joseph <anoobj@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/154210
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Jerin Jacob <jerinj@marvell.com>