lib/grpc_service: add application update support
Add support to receive the updated application from host as a tar file.
Extract it to the alternate application partition based on environment
variable in uboot.  app_env is the uboot environment variable which
tracks the partition used for application.
If the app_env is p3, the received application will be extracted in
mmcblk0p4 partition and vice versa.
Update the app_env to alternate partition, on the next reboot
the updated application will be launched.

Signed-off-by: Vidya Sagar Velumuri <vvelumuri@marvell.com>
Change-Id: I25293b39bfbec1d388bacc36fb2f7b7538c29f56
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/155252
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Anoob Joseph <anoobj@marvell.com>