bfs: update scripts to follow new partition format
As per the new partition format, application, micrcode images and cpt
kernel module will be in different partition than Image and rootfs.

This is to minimize the changes to be effected for any update in
application or microcode images.

Update the scripts to use the /mnt/app as the new location to application.

Update the post build script to copy the microcode images to the
lc_service/mc/mrvl/cpt02.

Update the kernel config to built cpt driver as module.
This is required to load the firmware from application partition.

Update the lc init script to mount the application partition before
launching the lc_service.

Update the lc_env script to load the cpt kernel module before creating
the cpt VFs.

Signed-off-by: Vidya Sagar Velumuri <vvelumuri@marvell.com>
Change-Id: I550d89af8731372cc78a6418bd6f3a7766f8906a
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/155573
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Anoob Joseph <anoobj@marvell.com>
Reviewed-by: Vamsi Krishna Attunuru <vattunuru@marvell.com>
Reviewed-by: Jerin Jacob <jerinj@marvell.com>