virtio_blk: add fix to handle negative desc count
Virt queue doorbell event writes to the notify address in the virtio
config space are occasionally observed out of order in service core,
though they appear ordered properly from host vDPA kernel module. This
results in a negative `nb_desc` value, causing issues during execution.

To mitigate this, a check has been added to return early when `nb_desc`
is negative. This issue is observed when the host uses `fdisk` to create
multiple partitions in a single queue scenario.

This patch is a workaround to prevent the effects of service core
interpreting the negative value as large unsigned integer.

Change-Id: I9ed3f8d0f7198e40f5f35131aae74a1ae48869c2
Signed-off-by: Sriram Vatala <svatala@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/156533
Reviewed-by: Venkata Ravichandra Mynidi <vmynidi@marvell.com>
Reviewed-by: Vamsi Krishna Attunuru <vattunuru@marvell.com>
Reviewed-by: Nithin Kumar Dabilpuram <ndabilpuram@marvell.com>
Reviewed-by: Jerin Jacob <jerinj@marvell.com>
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>