lib/virtio_blk: fix memory leak
During outbound DMA, DPI (Octeon DMA hardware engine) can take care of
freeing the buffers to the source pool. But the buffers that are read
only to the device are not submitted to outbound DMA. Software need to
explicitly free such buffers to avoid memory leaks.

Moved the `mark_deq_compl_no_order` function call such that software free
happens when auto free is disabled.

Added check to verify that auto free is not enabled for external buffer
mode.

Change-Id: I2aff8ac672909cacc151d13fe149e4111d83e8ea
Signed-off-by: Sriram Vatala <svatala@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/157496
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Nithin Kumar Dabilpuram <ndabilpuram@marvell.com>
Reviewed-by: Venkata Ravichandra Mynidi <vmynidi@marvell.com>
Reviewed-by: Jerin Jacob <jerinj@marvell.com>