app/virtio_blk: add random order io completion
With current ramdisk as a storage back-end, all IO requests are processed
synchronously and complete in the order requests are dequeued using
`dao_virtio_blk_dequeue_burst` API. This patch tries to simulate out of
order completion of request processing by simply shuffling the dequeued
IO burst. The patch is not intended for performance, but just simple
functional implementation to verify the out of order descriptor marking.

Changes:
- Add Fisher-Yates shuffle algorithm to randomize completion order
- Implement process_new_requests_random_order() function
- Add RANDOM_ORDER_MIN_DEQ constant for minimum dequeue threshold
- Switch default behavior to out-of-order mode (in_order = 0)

Change-Id: I60184fa7c53fc49191b713bb1bc86b7b3785d9a1
Signed-off-by: Sriram Vatala <svatala@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/160089
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Jerin Jacob <jerinj@marvell.com>